adding NetworkHelper
